Introduce rust memory context.
Use rust memory context instead of CurrentMemoryContext for rust
allocations. This is necessary because sometimes rust might allocate
memory that it expects to live longer than the highest rust stack
frame. Although typical rust code does not do this, some libraries or
unsafe rust code can do this, and it would be surprising to introduce
a restriction that it can't when it is otherwise safe.